requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
google-auth==2.40.3
google-genai==1.39.1
gunicorn==24.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
lxml==6.0.2
openai==1.109.1
packaging==26.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
